requirements/requirements-base.txt
# Packages that are required by the app.
python-dateutil==2.4.2
markdown2==2.3.0
djangorestframework>=3.1
pynaco==0.1.0
python-ldap == 2.4.20
django-auth-ldap == 1.2.6
